backend/requirements.txt
fastapi
uvicorn
pydantic
torch
pillow
python-multipart
aiofiles
python-dotenv
numpy
opencv-python
matplotlib